fuzz: move patcher engine into ragu_testing
