+++ crates/ragu_testing/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod circuits;
 pub mod pcd;
+pub mod recorder;
 pub mod registry;
 pub mod strategies;
 pub mod substrate;
